requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
openai>=1.59.0,<2
python-dotenv>=1.0.0,<2
google-genai>=1.0.0,<2
